This project is done
💻EXERCISES:
 1) When pressing on Load Images button, load the pictures from
https://api.pexels.com/v1/search?query=[your-query]   2) When pressing on Load Seconday Images, load the pictures from
https://api.pexels.com/v1/search?query=[your-secondary-query]

3) The Edit button should be replaced with a "Hide" button.

4) When the hide button is pressed, the whole picture card should disappear.

5) Replace the "9 mins" string in the card template with the ID of the Image    6) Add in the "jumbotron" a search field. Use the value of the search field to search new images
and replace the existing ones.

💻EXTRAS:
Handle API errors gracefully, using alert components with a custom message inside
Add a carousel at the bottom of the page with images from another API call
When the user clicks on the view button inside a card, open that image in a modal
💻MORE EXTRAS:
Use the .map method to create, from your API response, an array containing just the url strings (you can display the result with a console.log)
Use the .filter method to modify the result of the API call. Keep only the images from a specific author or your chose (you can display the result with a console.log)